backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pywebpush>=2.0.0
apscheduler>=3.10.0
httpx>=0.27.0
